* 修复 右键菜单抢焦点问题 * 修改一些llvm的警告 * 修复上个版本导致devtools显示不了高亮节点的问题 * 修复layui demo快速点击导致网络层崩溃的问题

+++ content/ui/ToolTip.h
@@ -158,7 +158,7 @@ private:
 
     bool isNearPos(const POINT& a, const POINT& b)
     {
-        return std::abs(a.x - b.x) + std::abs(a.y - b.y) < 15;
+        return std::abs((int)(a.x - b.x)) + std::abs((int)(a.y - b.y)) < 15;
     }
 
     void delayShowTimerFired(blink::Timer<ToolTip>*)
